+++ src/js/api/order.js
@@ -280,6 +280,14 @@ ripe.Ripe.prototype.createNoteOrderP = function(number, text, options) {
     });
 };
 
+/**
+ * Creates a shipping waybill for the order with the provided number.
+ *
+ * @param {Number} number The number of the order to create the waybill for.
+ * @param {Object} options An object of options to configure the request.
+ * @param {Function} callback Function with the result of the request.
+ * @returns {XMLHttpRequest} The XMLHttpRequest instance of the API request.
+ */
 ripe.Ripe.prototype.createWaybillOrder = function(number, options, callback) {
     callback = typeof options === "function" ? options : callback;
     options = typeof options === "function" || options === undefined ? {} : options;
@@ -293,9 +301,88 @@ ripe.Ripe.prototype.createWaybillOrder = function(number, options, callback) {
     return this._cacheURL(options.url, options, callback);
 };
 
+/**
+ * Creates a shipping waybill for the order with the provided number.
+ *
+ * @param {Number} number The number of the order to create the waybill for.
+ * @param {Object} options An object of options to configure the request.
+ * @returns {Promise} The contents of the note instance that was created.
+ */
 ripe.Ripe.prototype.createWaybillOrderP = function(number, options) {
     return new Promise((resolve, reject) => {
         this.createWaybillOrder(number, options, (result, isValid, request) => {
+            isValid ? resolve(result) : reject(new ripe.RemoteError(request, null, result));
+        });
+    });
+};
+
+/**
+ * Creates a shipping return waybill for the order with the provided number.
+ *
+ * @param {Number} number The number of the order to create the return waybill for.
+ * @param {Object} options An object of options to configure the request.
+ * @param {Function} callback Function with the result of the request.
+ * @returns {XMLHttpRequest} The XMLHttpRequest instance of the API request.
+ */
+ripe.Ripe.prototype.createReturnWaybillOrder = function(number, options, callback) {
+    callback = typeof options === "function" ? options : callback;
+    options = typeof options === "function" || options === undefined ? {} : options;
+    const url = `${this.url}orders/${number}/return_waybill`;
+    options = Object.assign(options, {
+        url: url,
+        method: "POST",
+        auth: true
+    });
+    options = this._build(options);
+    return this._cacheURL(options.url, options, callback);
+};
+
+/**
+ * Creates a shipping return waybill for the order with the provided number.
+ *
+ * @param {Number} number The number of the order to create the return waybill for.
+ * @param {Object} options An object of options to configure the request.
+ * @returns {Promise} The contents of the note instance that was created.
+ */
+ripe.Ripe.prototype.createReturnWaybillOrderP = function(number, options) {
+    return new Promise((resolve, reject) => {
+        this.createReturnWaybillOrder(number, options, (result, isValid, request) => {
+            isValid ? resolve(result) : reject(new ripe.RemoteError(request, null, result));
+        });
+    });
+};
+
+/**
+ * Manually trigger an order shipping info refresh.
+ *
+ * @param {Number} number The number of the order to refresh the shipping info.
+ * @param {Object} options An object of options to configure the request.
+ * @param {Function} callback Function with the result of the request.
+ * @returns {XMLHttpRequest} The XMLHttpRequest instance of the API request.
+ */
+ripe.Ripe.prototype.refreshShippingOrder = function(number, options, callback) {
+    callback = typeof options === "function" ? options : callback;
+    options = typeof options === "function" || options === undefined ? {} : options;
+    const url = `${this.url}orders/${number}/refresh_shipping`;
+    options = Object.assign(options, {
+        url: url,
+        method: "POST",
+        auth: true
+    });
+    options = this._build(options);
+    return this._cacheURL(options.url, options, callback);
+};
+
+/**
+ * Manually trigger an order shipping info refresh.
+ *
+ * @param {Number} number The number of the order to refresh the shipping info.
+ * @param {Object} options An object of options to configure the request.
+ * @returns {Promise} The order requested by number.
+ */
+ripe.Ripe.prototype.refreshShippingOrderP = function(number, options) {
+    return new Promise((resolve, reject) => {
+        this.refreshShippingOrder(number, options, (result, isValid, request) => {
             isValid ? resolve(result) : reject(new ripe.RemoteError(request, null, result));
         });
     });
@@ -747,6 +834,26 @@ ripe.Ripe.prototype.sendOrderP = function(number, trackingNumber, trackingUrl, o
 };
 
 /**
+ * Sets the order status to 'blocked'.
+ *
+ * @param {Number} number The number of the order to update.
+ * @param {Object} options An object of options to configure the request.
+ * @param {Function} callback Function with the result of the request.
+ * @returns {XMLHttpRequest} The XMLHttpRequest instance of the API request.
+ */
+ripe.Ripe.prototype.blockOrder = function(number, options, callback) {
+    return this.setOrderStatus(number, "block", options, callback);
+};
+
+ripe.Ripe.prototype.blockOrderP = function(number, options) {
+    return new Promise((resolve, reject) => {
+        this.blockOrder(number, options, (result, isValid, request) => {
+            isValid ? resolve(result) : reject(new ripe.RemoteError(request, null, result));
+        });
+    });
+};
+
+/**
  * Sets the order status to 'receive'.
  *
  * @param {Number} number The number of the order to update.
@@ -950,6 +1057,108 @@ ripe.Ripe.prototype.setTrackingP = function(number, trackingNumber, trackingUrl,
                 isValid ? resolve(result) : reject(new ripe.RemoteError(request, null, result));
             }
         );
+    });
+};
+
+/**
+ * Changes the return tracking info of an order.
+ *
+ * @param {Number} number The number of the order to change the return tracking info.
+ * @param {String} returnTrackingNumber The new return tracking number.
+ * @param {String} returnTrackingUrl The new return tracking URL.
+ * @param {Object} options An object of options to configure the request.
+ * @param {Function} callback Function with the result of the request.
+ * @returns {XMLHttpRequest} The XMLHttpRequest instance of the API request.
+ */
+ripe.Ripe.prototype.setReturnTracking = function(
+    number,
+    returnTrackingNumber,
+    returnTrackingUrl,
+    options,
+    callback
+) {
+    callback = typeof options === "function" ? options : callback;
+    options = typeof options === "function" || options === undefined ? {} : options;
+    const url = `${this.url}orders/${number}/return_tracking`;
+    options = Object.assign(options, {
+        url: url,
+        method: "PUT",
+        auth: true,
+        params: {
+            return_tracking_number: returnTrackingNumber,
+            return_tracking_url: returnTrackingUrl
+        }
+    });
+    options = this._build(options);
+    return this._cacheURL(options.url, options, callback);
+};
+
+/**
+ * Changes the return tracking info of an order.
+ *
+ * @param {Number} number The number of the order to change the return tracking info.
+ * @param {String} returnTrackingNumber The new return tracking number.
+ * @param {String} returnTrackingUrl The new return tracking URL.
+ * @param {Object} options An object of options to configure the request.
+ * @returns {Promise} The result of the order tracking info change.
+ */
+ripe.Ripe.prototype.setReturnTrackingP = function(
+    number,
+    returnTrackingNumber,
+    returnTrackingUrl,
+    options
+) {
+    return new Promise((resolve, reject) => {
+        this.setReturnTracking(
+            number,
+            returnTrackingNumber,
+            returnTrackingUrl,
+            options,
+            (result, isValid, request) => {
+                isValid ? resolve(result) : reject(new ripe.RemoteError(request, null, result));
+            }
+        );
+    });
+};
+
+/**
+ * Changes the proof of delivery info of an order.
+ *
+ * @param {Number} number The number of the order to change the proof of delivery info.
+ * @param {String} proofOfDeliveryUrl The new proof of delivery URL.
+ * @param {Object} options An object of options to configure the request.
+ * @param {Function} callback Function with the result of the request.
+ * @returns {XMLHttpRequest} The XMLHttpRequest instance of the API request.
+ */
+ripe.Ripe.prototype.setProofOfDelivery = function(number, proofOfDeliveryUrl, options, callback) {
+    callback = typeof options === "function" ? options : callback;
+    options = typeof options === "function" || options === undefined ? {} : options;
+    const url = `${this.url}orders/${number}/proof_of_delivery`;
+    options = Object.assign(options, {
+        url: url,
+        method: "PUT",
+        auth: true,
+        params: {
+            proof_of_delivery_url: proofOfDeliveryUrl
+        }
+    });
+    options = this._build(options);
+    return this._cacheURL(options.url, options, callback);
+};
+
+/**
+ * Changes the proof of delivery info of an order.
+ *
+ * @param {Number} number The number of the order to change the proof of delivery info.
+ * @param {String} proofOfDeliveryUrl The new proof of delivery URL.
+ * @param {Object} options An object of options to configure the request.
+ * @returns {Promise} The result of the order proof of delivery info change.
+ */
+ripe.Ripe.prototype.setProofOfDeliveryP = function(number, proofOfDeliveryUrl, options) {
+    return new Promise((resolve, reject) => {
+        this.setProofOfDelivery(number, proofOfDeliveryUrl, options, (result, isValid, request) => {
+            isValid ? resolve(result) : reject(new ripe.RemoteError(request, null, result));
+        });
     });
 };
 
@@ -1189,6 +1398,41 @@ ripe.Ripe.prototype.unsubscribeOrderP = function(number, options) {
 };
 
 /**
+ * Marks an order as having been touched by the user that
+ * is currently authenticated.
+ *
+ * @param {Number} number The number of the order to touch.
+ * @returns {XMLHttpRequest} The order.
+ */
+ripe.Ripe.prototype.touchOrder = function(number, options, callback) {
+    callback = typeof options === "function" ? options : callback;
+    options = typeof options === "function" || options === undefined ? {} : options;
+    const url = `${this.url}orders/${number}/touch`;
+    options = Object.assign(options, {
+        url: url,
+        auth: true,
+        method: "PUT"
+    });
+    options = this._build(options);
+    return this._cacheURL(options.url, options, callback);
+};
+
+/**
+ * Marks an order as having been touched by the user that
+ * is currently authenticated.
+ *
+ * @param {Number} number The number of the order to touch.
+ * @returns {Promise} The order.
+ */
+ripe.Ripe.prototype.touchOrderP = function(number, options) {
+    return new Promise((resolve, reject) => {
+        this.touchOrder(number, options, (result, isValid, request) => {
+            isValid ? resolve(result) : reject(new ripe.RemoteError(request, null, result));
+        });
+    });
+};
+
+/**
  * @ignore
  */
 ripe.Ripe.prototype._getOrderReportURL = function(number, key, options) {
@@ -1245,6 +1489,7 @@ ripe.Ripe.prototype._importOrder = function(ffOrderId, options = {}) {
             ? this.initialsExtra
             : options.initialsExtra || options.initials_extra;
     const gender = options.gender === undefined ? null : options.gender;
+    const scale = options.scale === undefined ? null : options.scale;
     const size = options.size === undefined ? null : options.size;
     const pending = options.pending === undefined ? null : options.pending;
     const notify = options.notify === undefined ? null : options.notify;
@@ -1274,6 +1519,8 @@ ripe.Ripe.prototype._importOrder = function(ffOrderId, options = {}) {
     if (variant) contents.variant = variant;
     if (productId) contents.product_id = productId;
     if (gender) contents.gender = gender;
+    if (scale) contents.scale = scale;
+    if (price) contents.price = price;
     if (Object.keys(initialsExtra).length > 0) {
         contents.initials_extra = initialsExtra;
     } else if (initials && engraving) {
@@ -1291,7 +1538,6 @@ ripe.Ripe.prototype._importOrder = function(ffOrderId, options = {}) {
     };
     if (type) params.type = type;
     if (country) params.country = country;
-    if (price) params.price = price;
     if (currency) params.currency = currency;
     if (meta) params.meta = meta;
     if (notes) params.notes = notes;